+++ pax/MongoManager.py
@@ -19,10 +19,11 @@ class MongoManager:
         self.monary_enabled = MONARY_ENABLED
 
     def get_database(self, database_name=None, uri=None, monary=False):
-        """Get a connection to the database_name. Returns Mongo database access object.
+        """Get a connection to the database_name. Returns Mongo database object.
         If you provide a mongodb connection string uri, we will insert user & password into it,
         otherwise one will be built from the configuration settings.
-        If database_name=None, will connect to the default database of the uri.
+        If database_name=None, will connect to the default database of the uri. database=something
+        overrides event the uri's specification of a database.
         If there is already a client connecting to this database, no new mongoclient will be created.
         """
         # Pattern of URI's we expect from database (without user & pass)
@@ -38,7 +39,9 @@ class MongoManager:
             m = re.match(uri_pattern, uri)
             if m:
                 # URI was provided, but without user & pass.
-                host, port, database_name = m.groups()
+                host, port, _database_name = m.groups()
+                if database_name is None:
+                    database_name = _database_name
                 uri = full_uri_format.format(database=database_name, host=host, port=port,
                                              user=self.config['user'], password=self.config['password'])
             else: